Split public views
