+++ statistics/tests/test_highscores.py
@@ -1,0 +1,11 @@
+from django.test import TestCase
+from django.core.urlresolvers import reverse
+
+from statistics.views import changelog
+
+
+class HighscoresTests(TestCase):
+
+    def test_highscores_loads(self):
+        response = self.client.get(reverse("statistics:high_scores"))
+        self.assertEqual(response.status_code, 200)
