+++ statistics/tests/test_home.py
@@ -2,7 +2,7 @@ from django.test import TestCase
 from django.core.urlresolvers import reverse
 from django.contrib.auth.models import User
 
-from ..views.public_views import home
+from statistics.views import home
 from ..models import Session, Committee
 from helpers import create_session, create_committee, create_user_max
