+++ statistics/views/helpers.py
@@ -1,0 +1,4 @@
+from raven import Client
+from django.conf import settings
+
+raven_client = Client(settings.RAVEN_CONFIG['dsn'])
