+++ statistics/views/protected/__init__.py
