+++ statistics/views/public/get_started.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+
+def get_started(request):
+    return render(request, 'statistics/get_started.html')
